requirements.txt
Flask
mysql-connector-python
azure-storage-blob
python-dotenv
